requirements.txt
# IST AI Voice Agent - Dependencies
# Python 3.10+

# Web & API
flask>=3.0.0
gunicorn>=21.0.0
python-dotenv>=1.0.0

# LiveKit
livekit>=0.19.0
livekit-api>=0.8.0
livekit-agents>=1.0.0
livekit-plugins-groq>=0.8.0
livekit-plugins-silero>=0.9.0

# RAG & Embeddings
chromadb>=0.4.22
sentence-transformers>=2.2.0
rank-bm25>=0.2.2

# LLM & APIs
groq>=0.4.0

# TTS (Edge-TTS - no official LiveKit plugin; we use custom)
edge-tts>=6.1.0

# Data loading
pypdf>=3.0.0

# Utilities
httpx>=0.25.0
